Clean up docstrings and comments
I mostly wanted to remove the implied ordering of 'cw' and 'ccw' and drop the
references to the correspondance between direction and +/- frequency. This
correspondance can change depending on the hardware or agent configuration. In
the future this should be hidden by the agents so that a frequency and
direction are given. Then within sorunlib we can avoid needing to know the
correspondance internally.

Other changes are for readability and consistency with overall style within the
repo.