Add `wiregrid.time_constant` function (#179)
* Add time_constant function

* Add stream during insert/eject and sleep time before and after insert/eject

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Split stream before and after HWP is stopped

* Add try/finally to make it sure to streams are stopped

* Add `_get_direction()` function to HWP module

* Change hwp direction label from `forward`/`backward` to `ccw`/`cw`.

* Clean up docstrings and comments

I mostly wanted to remove the implied ordering of 'cw' and 'ccw' and drop the
references to the correspondence between direction and +/- frequency. This
correspondence can change depending on the hardware or agent configuration. In
the future this should be hidden by the agents so that a frequency and
direction are given. Then within sorunlib we can avoid needing to know the
correspondence internally.

Other changes are for readability and consistency with overall style within the
repo.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Shunsuke Adachi <38268229+sadachi5@users.noreply.github.com>
Co-authored-by: Brian Koopman <BrianJKoopman@users.noreply.github.com>