Reduce check_started timeout
This increased timeout came from #147. Back then, the session.status was
manually set to running, but only after the call to 'stream_g3_on' in the
agent. This meant that check_started would actually block until the stream was
enabled. Now that ocs automatically sets the status to running, this no longer
works, which is why we need this additional check.

check_started should return quickly now, and shouldn't come close to this 60
second timeout.